Test method for name checking added
